+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Enhancing and refactoring of `ExtCoSimulation` [#237](https://github.com/ie3-institute/simonaAPI/issues/237)
+- Converted `Optional<Long>` to `OptionalLong` [#451](https://github.com/ie3-institute/simonaAPI/issues/451)
 
 ### Fixed